CROP

+++ alocari.c
@@ -5,9 +5,9 @@
 #include "alocari.h"
 
 // TODO
-void eliberare_matrice_pixeli(union pixel **mat, size_t n)
+void eliberare_matrice_pixeli(union pixel **mat, long n)
 {
-	for (size_t i = 0; i < n; ++i)
+	for (long i = 0; i < n; ++i)
 		free(mat[i]);
 	free(mat);
 }
@@ -18,13 +18,13 @@ void eliberare_imagine(struct imagine *img)
 	free(img);
 }
 
-union pixel **aloca_matrice_pixeli(size_t n, size_t m)
+union pixel **aloca_matrice_pixeli(long n, long m)
 {
 	union pixel **mat = (union pixel **)malloc(n * sizeof(union pixel *));
 	if (!mat)
 		return NULL;
 
-	for (size_t i = 0; i < n; ++i) {
+	for (long i = 0; i < n; ++i) {
 		mat[i] = (union pixel *)malloc(m * sizeof(union pixel));
 		if (!mat[i]) {
 			eliberare_matrice_pixeli(mat, i);
