+++ fisiere.c
@@ -11,8 +11,8 @@
 static int verif_format(FILE *f);
 // Returneaza prima linie care nu contine un comentariu (nu incepe cu #).
 static char *omite_comentarii(FILE *f);
-static unsigned char *citeste_valori_ascii(FILE *f, size_t n);
-static unsigned char *citeste_valori_binar(FILE *f, size_t n);
+static unsigned char *citeste_valori_ascii(FILE *f, long n);
+static unsigned char *citeste_valori_binar(FILE *f, long n);
 
 char *citire_linie(FILE *stream)
 {
@@ -73,8 +73,8 @@ struct imagine *incarcare_fisier()
 
 	// TODO
 	char *end;
-	size_t latime = strtoll(linie, &end, 10);
-	size_t inaltime = strtoll(end, NULL, 10);
+	long latime = strtoll(linie, &end, 10);
+	long inaltime = strtoll(end, NULL, 10);
 	int val_max = 1; // TODO
 
 	// Imaginile alb-negru nu au si o valoare maxima a culorii.
@@ -104,8 +104,8 @@ struct imagine *incarcare_fisier()
 	img->inaltime = inaltime;
 	img->latime = latime;
 
-	for (size_t i = 0; i < inaltime; ++i) {
-		for (size_t j = 0; j < latime; ++j) {
+	for (long i = 0; i < inaltime; ++i) {
+		for (long j = 0; j < latime; ++j) {
 			union pixel *p = &img->pixeli[i][j];
 			if (color) {
 				p->culoare.r = a[i * latime + j * octeti_per_pixel];
@@ -153,8 +153,8 @@ void salvare_imagine(struct imagine img)
 	if (format % 3 != 1)
 		fprintf(f, "%d\n", img.val_max);
 
-	for (size_t i = 0; i < img.inaltime; ++i) {
-		for (size_t j = 0; j < img.latime; ++j) {
+	for (long i = 0; i < img.inaltime; ++i) {
+		for (long j = 0; j < img.latime; ++j) {
 			union pixel p = img.pixeli[i][j];
 			if (format < 4) {
 				if (img.color)
@@ -200,7 +200,7 @@ static char *omite_comentarii(FILE *f)
 	return linie;
 }
 
-static unsigned char *citeste_valori_ascii(FILE *f, size_t n)
+static unsigned char *citeste_valori_ascii(FILE *f, long n)
 {
 	unsigned char *v = (unsigned char *)malloc(n * sizeof(unsigned char));
 	if (!v)
@@ -210,7 +210,7 @@ static unsigned char *citeste_valori_ascii(FILE *f, size_t n)
 	char *p = linie;
 	char *q;
 
-	size_t i = 0;
+	long i = 0;
 	while (i < n) {
 		unsigned char val = (unsigned char)strtol(p, &q, 10);
 		// Nu mai exista valori pe rand, se trece pe urmatorul.
@@ -228,7 +228,7 @@ static unsigned char *citeste_valori_ascii(FILE *f, size_t n)
 	return v;
 }
 
-static unsigned char *citeste_valori_binar(FILE *f, size_t n)
+static unsigned char *citeste_valori_binar(FILE *f, long n)
 {
 	unsigned char *v = (unsigned char *)malloc(n * sizeof(unsigned char));
 	if (!v)
